implemented semesters for work and SWS, replaced float with double in some structures, did not test yet.

+++ src/main/java/com/example/profisbackend/dto/studyprogram/StudyProgramDTO.java
@@ -2,6 +2,6 @@ package com.example.profisbackend.dto.studyprogram;
 
 import com.example.profisbackend.enums.DegreeType;
 
-public record StudyProgramDTO(DegreeType degreeType, String title, float sws) {
+public record StudyProgramDTO(DegreeType degreeType, String title, double sws) {
 }
