+++ src/main/java/com/example/profisbackend/mapper/MarkMapper.java
@@ -1,7 +1,7 @@
 package com.example.profisbackend.mapper;
 
 public class MarkMapper {
-    public static String pointsToGermanNote(Float points) throws IllegalArgumentException {
+    public static String pointsToGermanNote(double points) throws IllegalArgumentException {
         if (points < 0 || points > 100) {
             throw new IllegalArgumentException();
         }
@@ -11,11 +11,11 @@ public class MarkMapper {
         if (points < 50) {
             return "5,0";
         }
-        Float mark = 1.0F;
-        Float n = 95F;
+        double mark = 1.0F;
+        double n = 95F;
         while (points < n) {
-            n = (float) (n - 1.5);
-            mark = (float) (mark + 0.1);
+            n = (double) (n - 1.5);
+            mark = (double) (mark + 0.1);
         }
         return String.format("%.1f", mark).replace('.', ',');
     }
